indexer/tests/confesster.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { BigInt, Address } from "@graphprotocol/graph-ts"
import { CoffeeBought } from "../generated/schema"
import { CoffeeBought as CoffeeBoughtEvent } from "../generated/Confesster/Confesster"
import { handleCoffeeBought } from "../src/confesster"
import { createCoffeeBoughtEvent } from "./confesster-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let id = BigInt.fromI32(234)
    let buyer = Address.fromString("0x0000000000000000000000000000000000000001")
    let author = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let message = "Example string value"
    let timestamp = BigInt.fromI32(234)
    let newCoffeeBoughtEvent = createCoffeeBoughtEvent(
      id,
      buyer,
      author,
      message,
      timestamp
    )
    handleCoffeeBought(newCoffeeBoughtEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("CoffeeBought created and stored", () => {
    assert.entityCount("CoffeeBought", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "CoffeeBought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "buyer",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "CoffeeBought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "author",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "CoffeeBought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "message",
      "Example string value"
    )
    assert.fieldEquals(
      "CoffeeBought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "timestamp",
      "234"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
